refactor: 适配远程users表结构，移除多余字段，id改为Integer
- User实体映射wallet_address_id列，移除nickname/avatar/email/phone
- id类型Long改为Integer，移除FieldFill策略依赖数据库默认值
- JWT及全链路userId类型同步为Integer
- schema.sql同步更新表结构

+++ src/main/java/com/gw/server/security/JwtAuthFilter.java
@@ -37,7 +37,7 @@ public class JwtAuthFilter implements HandlerInterceptor {
         }
 
         // Store userId and address in request attributes for controllers to use
-        Long userId = jwtTokenProvider.getUserIdFromToken(token);
+        Integer userId = jwtTokenProvider.getUserIdFromToken(token);
         String address = jwtTokenProvider.getAddressFromToken(token);
         request.setAttribute("userId", userId);
         request.setAttribute("walletAddress", address);
